Move readUser code outside of the file package
